Move to subpackage and show only names of DBs in ResultWindow
