+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/identification/sirius/ResultWindow.java
@@ -41,12 +41,14 @@ import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
-import javax.swing.ListSelectionModel;
 import javax.swing.SwingUtilities;
 
 import net.sf.mzmine.datamodel.PeakListRow;
 import net.sf.mzmine.main.MZmineCore;
 import net.sf.mzmine.modules.peaklistmethods.identification.sirius.db.DBFrame;
+import net.sf.mzmine.modules.peaklistmethods.identification.sirius.table.ResultTable;
+import net.sf.mzmine.modules.peaklistmethods.identification.sirius.table.ResultTableModel;
+import net.sf.mzmine.modules.peaklistmethods.identification.sirius.table.SiriusCompound;
 import net.sf.mzmine.taskcontrol.Task;
 import net.sf.mzmine.taskcontrol.TaskStatus;
 import net.sf.mzmine.util.GUIUtils;
