+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/identification/sirius/db/DBFrame.java
@@ -38,7 +38,7 @@ import javax.swing.ListSelectionModel;
 import javax.swing.SwingUtilities;
 import javax.swing.table.TableRowSorter;
 import net.sf.mzmine.main.MZmineCore;
-import net.sf.mzmine.modules.peaklistmethods.identification.sirius.SiriusCompound;
+import net.sf.mzmine.modules.peaklistmethods.identification.sirius.table.SiriusCompound;
 import net.sf.mzmine.util.GUIUtils;
 
 public class DBFrame extends JFrame implements ActionListener {
