+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/identification/sirius/table/ResultTableModel.java
@@ -17,7 +17,7 @@
  * Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-package net.sf.mzmine.modules.peaklistmethods.identification.sirius;
+package net.sf.mzmine.modules.peaklistmethods.identification.sirius.table;
 
 import java.text.NumberFormat;
 import java.util.Vector;
